refactor(frontend): split 1298-line report-package page into section components
Extract the inline report-body JSX from reports/package/page.tsx into
co-located presentational components under components/reports/package/:
- PackageChrome (cover, TOC, setup, skeleton, framework selection)
- PackageSnapshotsCard
- PackageReadinessSections (readiness, source trust, framework policy)
- PackageScheduleSections (section cards, annualized, notes, export contract)
- PackageTraceability (traceability table + lineage modal)
- shared.tsx (constants, types, pure formatters)

page.tsx now holds only state, the data hook, action handlers, and
orchestration: 1298 -> 293 lines. Rendered markup is byte-identical, so the
existing personalReportPackagePage test (1585 lines) passes unchanged.

Components live under src/components (not an app/ _components dir) so the
Next client-entry serializable-props rule does not false-positive on the
callback props.

typecheck + lint + 826 FE tests green.
